+++ src/backend/InvenTree/order/tasks.py
@@ -1,7 +1,6 @@
 """Background tasks for the 'order' app."""
 
 from datetime import datetime, timedelta
-from typing import Union
 
 from django.contrib.auth.models import Group, User
 from django.db import transaction
@@ -105,7 +104,7 @@ def check_overdue_purchase_orders():
 @tracer.start_as_current_span('notify_overdue_sales_order')
 def notify_overdue_sales_order(so: order.models.SalesOrder) -> None:
     """Notify appropriate users that a SalesOrder has just become 'overdue'."""
-    targets: list[Union[User, Group, Owner]] = []
+    targets: list[User | Group | Owner] = []
 
     if so.created_by:
         targets.append(so.created_by)
@@ -172,7 +171,7 @@ def check_overdue_sales_orders():
 @tracer.start_as_current_span('notify_overdue_return_order')
 def notify_overdue_return_order(ro: order.models.ReturnOrder) -> None:
     """Notify appropriate users that a ReturnOrder has just become 'overdue'."""
-    targets: list[Union[User, Group, Owner]] = []
+    targets: list[User | Group | Owner] = []
 
     if ro.created_by:
         targets.append(ro.created_by)