+++ src/backend/InvenTree/plugin/base/barcodes/mixins.py
@@ -2,8 +2,6 @@
 
 from __future__ import annotations
 
-from typing import Optional
-
 from django.core.exceptions import ValidationError
 from django.db.models import Q
 from django.utils.translation import gettext_lazy as _
@@ -115,7 +113,7 @@ class SupplierBarcodeMixin(BarcodeMixin):
 
         return fields.get(key, backup_value)
 
-    def get_part(self) -> Optional[Part]:
+    def get_part(self) -> Part | None:
         """Extract the Part object from the barcode fields."""
         # TODO: Implement this
         return None
@@ -130,7 +128,7 @@ class SupplierBarcodeMixin(BarcodeMixin):
         """Return the supplier part number from the barcode fields."""
         return self.get_field_value(self.SUPPLIER_PART_NUMBER)
 
-    def get_supplier_part(self) -> Optional[SupplierPart]:
+    def get_supplier_part(self) -> SupplierPart | None:
         """Return the SupplierPart object for the scanned barcode.
 
         Returns:
@@ -174,7 +172,7 @@ class SupplierBarcodeMixin(BarcodeMixin):
         """Return the manufacturer part number from the barcode fields."""
         return self.get_field_value(self.MANUFACTURER_PART_NUMBER)
 
-    def get_manufacturer_part(self) -> Optional[ManufacturerPart]:
+    def get_manufacturer_part(self) -> ManufacturerPart | None:
         """Return the ManufacturerPart object for the scanned barcode.
 
         Returns:
@@ -215,7 +213,7 @@ class SupplierBarcodeMixin(BarcodeMixin):
         """Return the supplier order number from the barcode fields."""
         return self.get_field_value(self.SUPPLIER_ORDER_NUMBER)
 
-    def get_purchase_order(self) -> Optional[PurchaseOrder]:
+    def get_purchase_order(self) -> PurchaseOrder | None:
         """Extract the PurchaseOrder object from the barcode fields.
 
         Inspect the customer_order_number and supplier_order_number fields,
@@ -262,7 +260,7 @@ class SupplierBarcodeMixin(BarcodeMixin):
             'extract_barcode_fields must be implemented by each plugin'
         )
 
-    def scan(self, barcode_data: str) -> Optional[dict]:
+    def scan(self, barcode_data: str) -> dict | None:
         """Perform a generic 'scan' operation on a supplier barcode.
 
         The supplier barcode may provide sufficient information to match against
@@ -321,7 +319,7 @@ class SupplierBarcodeMixin(BarcodeMixin):
         location=None,
         auto_allocate: bool = True,
         **kwargs,
-    ) -> Optional[dict]:
+    ) -> dict | None:
         """Attempt to receive an item against a PurchaseOrder via barcode scanning.
 
         Arguments:
@@ -432,7 +430,7 @@ class SupplierBarcodeMixin(BarcodeMixin):
 
         return response
 
-    def get_supplier(self, cache: bool = False) -> Optional[Company]:
+    def get_supplier(self, cache: bool = False) -> Company | None:
         """Get the supplier for the SUPPLIER_ID set in the plugin settings.
 
         If it's not defined, try to guess it and set it if possible.