+++ src/backend/InvenTree/stock/models.py
@@ -5,7 +5,6 @@ from __future__ import annotations
 import os
 from datetime import timedelta
 from decimal import Decimal, InvalidOperation
-from typing import Optional
 
 from django.conf import settings
 from django.contrib.auth.models import User
@@ -129,7 +128,7 @@ class StockLocationReportContext(report.mixins.BaseReportContext):
 
     location: StockLocation
     qr_data: str
-    parent: Optional[StockLocation]
+    parent: StockLocation | None
     stock_location: StockLocation
     stock_items: report.mixins.QuerySet[StockItem]
 
@@ -392,7 +391,7 @@ class StockItemReportContext(report.mixins.BaseReportContext):
     barcode_hash: str
     batch: str
     child_items: report.mixins.QuerySet[StockItem]
-    ipn: Optional[str]
+    ipn: str | None
     installed_items: set[StockItem]
     item: StockItem
     name: str
@@ -403,7 +402,7 @@ class StockItemReportContext(report.mixins.BaseReportContext):
     quantity: Decimal
     result_list: list[StockItemTestResult]
     results: dict[str, StockItemTestResult]
-    serial: Optional[str]
+    serial: str | None
     stock_item: StockItem
     tests: dict[str, StockItemTestResult]
     test_keys: list[str]
@@ -667,7 +666,7 @@ class StockItem(
         return items
 
     @staticmethod
-    def convert_serial_to_int(serial: str) -> Optional[int]:
+    def convert_serial_to_int(serial: str) -> int | None:
         """Convert the provided serial number to an integer value.
 
         This function hooks into the plugin system to allow for custom serial number conversion.
@@ -1783,7 +1782,7 @@ class StockItem(
         self,
         entry_type: int,
         user: User,
-        deltas: Optional[dict] = None,
+        deltas: dict | None = None,
         notes: str = '',
         commit: bool = True,
         **kwargs,
@@ -1842,9 +1841,9 @@ class StockItem(
         self,
         quantity: int,
         serials: list[str],
-        user: Optional[User] = None,
-        notes: Optional[str] = '',
-        location: Optional[StockLocation] = None,
+        user: User | None = None,
+        notes: str | None = '',
+        location: StockLocation | None = None,
     ):
         """Split this stock item into unique serial numbers.
 
@@ -1951,7 +1950,7 @@ class StockItem(
             item.save()
 
     @transaction.atomic
-    def copyTestResultsFrom(self, other: StockItem, filters: Optional[dict] = None):
+    def copyTestResultsFrom(self, other: StockItem, filters: dict | None = None):
         """Copy all test results from another StockItem."""
         # Set default - see B006